LayoutTests/imported/w3c/web-platform-tests/import-maps/bare-specifiers.sub-expected.txt

CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
Blocked access to external URL https://www1.localhost:9443/import-maps/resources/log.js?pipe=sub&name=cross-origin-bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
Blocked access to external URL https://www1.localhost:9443/import-maps/resources/log.js?pipe=sub&name=cross-origin-bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
Blocked access to external URL https://www1.localhost:9443/import-maps/resources/log.js?pipe=sub&name=cross-origin-bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: TypeError: Module name, 'bare/to-bare' does not resolve to a valid URL.
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare/bare


PASS bare/bare: <script src type=module>
PASS bare/bare: <script src type=text/javascript>
PASS bare/bare: static import
PASS bare/bare: dynamic import (from module)
PASS bare/bare: dynamic import (from text/javascript)
PASS bare/cross-origin-bare: <script src type=module>
PASS bare/cross-origin-bare: <script src type=text/javascript>
FAIL bare/cross-origin-bare: static import assert_unreached: script's error event shouldn't be fired Reached unreachable code
FAIL bare/cross-origin-bare: dynamic import (from module) assert_unreached: dynamic import promise shouldn't be rejected Reached unreachable code
FAIL bare/cross-origin-bare: dynamic import (from text/javascript) assert_unreached: dynamic import promise shouldn't be rejected Reached unreachable code
PASS bare/to-data: <script src type=module>
PASS bare/to-data: <script src type=text/javascript>
PASS bare/to-data: static import
PASS bare/to-data: dynamic import (from module)
PASS bare/to-data: dynamic import (from text/javascript)
PASS bare/to-bare: <script src type=module>
PASS bare/to-bare: <script src type=text/javascript>
PASS bare/to-bare: static import
PASS bare/to-bare: dynamic import (from module)
PASS bare/to-bare: dynamic import (from text/javascript)
